Deduplicate exec-event heartbeat wake options
